Refactor batch CLI command module
